crawler/requirements.txt
httpx[http2]==0.27.*
trafilatura==1.12.*
feedparser==6.*
playwright==1.49.*
apscheduler==4.*
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
simhash==2.*
langdetect==1.0.*
pyyaml==6.*
sentence-transformers==3.*
numpy==1.*
pgvector==0.3.*